+++ java/src/test/java/com/ironcorelabs/ironcore_alloy_java/IroncoreAlloyTest.java
@@ -75,7 +75,7 @@ public class IroncoreAlloyTest {
                         new StandaloneSecret(2, new Secret(keyByteArray)),
                         new StandaloneSecret(1, new Secret(keyByteArray)))))));
         IroncoreAlloyTest.JavaHttpClient httpClient = new IroncoreAlloyTest.JavaHttpClient();
-        integrationSdk = new SaasShield(new SaasShieldConfiguration("http://localhost:32804", "0WUaXesNgbTAuLwn", 1.1f, httpClient, true));
+        integrationSdk = new SaasShield(new SaasShieldConfiguration("http://localhost:32804", "0WUaXesNgbTAuLwn", 1.1f, httpClient, false, true));
     } 
     
     @Test
@@ -445,7 +445,7 @@ public class IroncoreAlloyTest {
     @Test
     public void badConfigurationTest() throws AlloyException, ExecutionException, InterruptedException {
         IroncoreAlloyTest.JavaHttpClient httpClient = new IroncoreAlloyTest.JavaHttpClient();
-        var badSdk = new SaasShield(new SaasShieldConfiguration("https://bad-url", "0WUaXesNgbTAuLwn", 1.1f, httpClient, false));
+        var badSdk = new SaasShield(new SaasShieldConfiguration("https://bad-url", "0WUaXesNgbTAuLwn", 1.1f, httpClient, false, false));
         float[] data = new float[]{1.0f, 2.0f, 3.0f};
         PlaintextVector plaintext = new PlaintextVector(data, new SecretPath(""), new DerivationPath(""));
         var metadata = AlloyMetadata.newSimple(new TenantId("fake_tenant"));
@@ -461,7 +461,7 @@ public class IroncoreAlloyTest {
     public void httpNotAllowed() throws AlloyException, ExecutionException, InterruptedException {
         IroncoreAlloyTest.JavaHttpClient httpClient = new IroncoreAlloyTest.JavaHttpClient();
         AlloyException err = assertThrows(AlloyException.InvalidConfiguration.class, () ->{
-            new SaasShieldConfiguration("http://bad-url", "0WUaXesNgbTAuLwn", 1.1f, httpClient, false);
+            new SaasShieldConfiguration("http://bad-url", "0WUaXesNgbTAuLwn", 1.1f, httpClient, false, false);
         });
         assertTrue(err.getMessage().contains("insecure"));
     }