+++ src/standalone/standard.rs
@@ -161,8 +161,13 @@ impl StandaloneStandardClient {
             &self.config.secrets,
             &metadata.tenant_id,
         )?;
-        let encrypted_document: HashMap<_, _> =
-            encrypt_map(plaintext_document.document.0, self.rng.clone(), dek)?;
+        let encrypted_document: HashMap<_, _> = encrypt_map(
+            plaintext_document.document.0,
+            self.rng.clone(),
+            dek,
+            false,
+            None,
+        )?;
         Ok(EncryptedDocument {
             edek: plaintext_document.edek,
             document: encrypted_document,
@@ -293,13 +298,15 @@ impl StandardDocumentOps for StandaloneStandardClient {
     /// a format matching the encoding in the data store. z85/ascii85 users should first pass these bytes through
     /// `encode_prefix_z85` or `base85_prefix_padding`. Make sure you've read the documentation of those functions to
     /// avoid pitfalls when encoding across byte boundaries.
-    fn get_searchable_edek_prefix(&self, id: i32) -> Vec<u8> {
-        get_prefix_bytes_for_search(ironcore_documents::v5::key_id_header::KeyIdHeader::new(
-            self.get_edek_type(),
-            self.get_payload_type(),
-            KeyId(id as u32),
-        ))
-        .into()
+    fn get_searchable_edek_prefix(&self, id: i32) -> Result<Vec<u8>, AlloyError> {
+        Ok(
+            get_prefix_bytes_for_search(ironcore_documents::v5::key_id_header::KeyIdHeader::new(
+                self.get_edek_type(),
+                self.get_payload_type(),
+                KeyId(id as u32),
+            ))
+            .into(),
+        )
     }
 }
 
@@ -716,7 +723,7 @@ pub(crate) mod test {
     #[test]
     fn get_searchable_edek_prefix_works() {
         let client = default_client();
-        let result = client.get_searchable_edek_prefix(100);
+        let result = client.get_searchable_edek_prefix(100).unwrap();
         assert_eq!(result, vec![0, 0, 0, 100, 130, 0]);
     }
 }